feat: extract polling logic into internal/watch/ package
Moves core watch polling logic from internal/cmd/watch.go into a new
internal/watch/ package in the orchestration layer, making it reusable
by both godark watch and future daemon mode. The cmd layer is now a thin
wrapper that creates a Watch instance and calls Run().

- New Watch struct with New()/Run()/PollOnce()/HandleChangesRequested()/HandleApproved()
- Testability seams moved into the new package (drop watch prefix)
- internal/cmd/watch.go reduced to Cobra wiring + watch.New().Run(ctx)
- docs/architecture.json updated: internal/watch/ added to orchestration paths
- All existing watch tests migrated and pass

Closes #516
